Render one example per assistant turn; harden config, anonymizer, packaging
The whole-dialogue render trained every assistant turn but the last on a
prompt the server never builds: Qwen3 in non-thinking mode puts the empty
<think> block before the generation prompt and strips it from history. Each
assistant turn is now its own example whose prompt is the add_generation_prompt
render with dataset.chat_template_kwargs (default enable_thinking=false); the
prompt's tokens must be a prefix of the example's, content must appear
verbatim, otherwise TemplateCompatibilityError. About 3x tokens per epoch on
the example scenario, same trained tokens. The real Qwen3 template is vendored
under tests/templates and tested against.

- tcsft render: a dialogue exactly as the trainer sees it, loss spans marked
- runs write config.yaml, run.json and example.txt into output_dir
- config: unknown keys rejected, lr_scheduler validated, eval_batch_size,
  chat_template_kwargs; requested device must exist
- trainer refuses over-long dialogues by id instead of dropping them;
  kbit checkpointing follows the config; merge in fp32, bf16 once at the end
- schema: optional tool-call ids validated against pending calls; every call
  needs its result before the next turn
- anonymizer: whole-word name matching, camelCase/snake_case key equivalence
- generator: 22 out_of_scope topics (ceiling ~2060, named in the error),
  decline amounts overlap the happy path, tickets derived from dialogue text
- cli: transformers imported lazily (--help in 0.05s, bare install works),
  --config on stats/filter/render, per-branch breakdown from split
- tests: Qwen3 template tests, tiny-model training smoke (-m training);
  CI runs a CLI smoke on the light install and the training smoke on the full
- docs: per-turn rendering, remeasured demo (57m11s, eval loss 0.064),
  near-duplicate leakage measured instead of claimed, stale numbers fixed

+++ src/toolcall_sft/training/merge.py
@@ -10,9 +10,14 @@ __all__ = ["merge_adapter"]
 
 
 def merge_adapter(*, adapter_dir: Path, output_dir: Path) -> Path:
-    model = AutoPeftModelForCausalLM.from_pretrained(str(adapter_dir), dtype=torch.bfloat16)
+    # Merge in fp32 and round once at the end. The adapter delta is small next to the
+    # base weight; computing W + BA in bf16 rounds the delta before the sum and loses part
+    # of the update that training kept in fp32 for exactly this reason.
+    model = AutoPeftModelForCausalLM.from_pretrained(str(adapter_dir), dtype=torch.float32)
     merged = model.merge_and_unload()
+    merged = merged.to(torch.bfloat16)
     merged.save_pretrained(str(output_dir))
+    # The tokenizer saved next to the adapter carries the chat template training rendered with.
     tokenizer = AutoTokenizer.from_pretrained(str(adapter_dir))
     tokenizer.save_pretrained(str(output_dir))
     return output_dir
